reorg and filter context for chart Arranger query
