+++ charts/arranger.ts
@@ -1,0 +1,56 @@
+/*
+ *
+ * Copyright (c) 2025 The Ontario Institute for Cancer Research. All rights reserved
+ *
+ *  This program and the accompanying materials are made available under the terms of
+ *  the GNU Affero General Public License v3.0. You should have received a copy of the
+ *  GNU Affero General Public License along with this program.
+ *  If not, see <http://www.gnu.org/licenses/>.
+ *
+ *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
+ *  EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ *  OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT
+ *  SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ *  INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
+ *  TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+ *  OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
+ *  IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ *
+ */
+
+/**
+ * This entire file is a temporary solution to target Arranger v2 server
+ * It is consistent with usage across this repo
+ * TODO: Upgrade to use Arranger v3 in the project (server and client)
+ */
+
+import { gql, useQuery } from '@apollo/client';
+import useFiltersContext from 'components/pages/file-repository/hooks/useFiltersContext';
+
+/**
+ * TODO: update to use ArrangerV3 data providers
+ * TODO: NumericAgg
+ */
+const generateQuery = ({ field }) => gql`
+  query ChartsFileCentric($filters:JSON) {
+    file {
+      aggregations(filters: $filters) {
+        ${field} {
+          bucket_count
+          buckets {
+            doc_count
+            key
+          }
+        }
+      }
+    }
+  }
+`;
+
+export const useArrangerCharts = ({ field }) => {
+  const { filters } = useFiltersContext();
+  console.log('arranger', filters);
+  const query = generateQuery({ field });
+  return useQuery(query, { variables: { filters } });
+};
